fixed teacher controller to show approved teachers for users and all teachers for admins

+++ app/Http/Controllers/TeachersController.php
@@ -8,9 +8,10 @@ use App\Teacher;
 use App\User;
 use App\Http\Resources\Teacher as TeacherResource;
 use App\Http\Resources\Course as CourseResource;
-
+use App\Http\Middleware\CheckRole;
 class TeachersController extends Controller
 {
+   
     /**
      * Display a listing of the resource.
      *
@@ -18,7 +19,12 @@ class TeachersController extends Controller
      */
     public function index()
     {
-        $teachers=Teacher::paginate(10);
+        if(CheckRole::checkCurrentUserRoles("admin")){
+            $teachers=Teacher::paginate(10);
+        }else{
+            $teachers=Teacher::where("is_approved","1")-> paginate(10);
+        }
+       
 
         return TeacherResource::collection($teachers);
     }
@@ -33,14 +39,18 @@ class TeachersController extends Controller
      */
     public function store(Request $request)
     {
+        // get current user id
+        $user_id=auth()->user()->id;
+
         // verify user exists
-        $user=User::findOrFail($request->input('user_id'));
+        $user=User::findOrFail($user_id);
 
         
         // Check if the user is already a teacher
         if($user->teacher!=null){
             return abort(400,"the user is already a teacher");
         }
+
 
         $teacher=new Teacher;
         $teacher->user_id=$user->id;
@@ -66,8 +76,16 @@ class TeachersController extends Controller
     public function show($id)
     {
         $teacher=Teacher::findOrFail($id);
+        if($teacher->is_approved==0 && !CheckRole::checkCurrentUserRoles("admin") ){
+            return response()->json([
+                'message' => 'Teacher Not Found'], 404);
+        
+        }else{
+            return new TeacherResource($teacher);
+        }
+       
 
-        return new TeacherResource($teacher);
+        
     }
 
     
@@ -84,7 +102,7 @@ class TeachersController extends Controller
         $teacher->about_me=$request->input('about_me');
         
         // the Teacher is a approved by admin only
-        if(auth()->user()->is_admin==1){
+        if(CheckRole::checkCurrentUserRoles("admin")){
             $teacher->is_approved=input('is_approved');
         }
         
@@ -106,6 +124,13 @@ class TeachersController extends Controller
     public function destroy($id)
     {
         $teacher=Teacher::findOrFail($id);
+        // get current user id
+        $user_id=auth()->user()->id;
+        // the Teacher is a removed by admin only or the teacher himself
+        if($user_id!=$id && !CheckRole::checkCurrentUserRoles("admin")){
+            return response()->json([
+                'message' => 'Unauthorized'], 401);
+        }
         if($teacher->delete()){
             return new TeacherResource($teacher);
         }
@@ -122,7 +147,12 @@ class TeachersController extends Controller
      */
     public function showCoursesByTeacher($id)
     {
+        
         $teacher=Teacher::findOrFail($id);
+        if($teacher->is_approved==0){
+            return response()->json([
+                'message' => 'This teacher is not approved'], 403);
+        }
         $courses=$teacher->courses;
         return CourseResource::collection($courses);
     }
